Implement register renaming
Can now track renamed registers through the RMT and ROB, and determine
if an instruction is ready to execute.

Idea for this simulator is to store a pointer to the instruction in
the ROB for destinations, then the renamed source registers of later
instructions use that as a pointer to the "parent" instruction to be
able to check if it is ready or not.